+++ src/tests/lib_tests_parse_tests/catalog_and_container_option_tests.rs
@@ -292,6 +292,34 @@ fn parse_container_reset_keep_data_is_supported() {
             subcommand: ContainerSubcommand::Reset {
                 name: Some("web".to_owned()),
                 keep_data: true,
+                wipe_data: false,
+                yes: false,
+            },
+            repo_override: None,
+            output_json: true,
+        })
+    );
+}
+
+#[test]
+fn parse_container_reset_wipe_data_yes_is_supported() {
+    let cmd = parse_command(vec![
+        "container".to_owned(),
+        "web".to_owned(),
+        "reset".to_owned(),
+        "--wipe-data".to_owned(),
+        "--yes".to_owned(),
+        "--json".to_owned(),
+    ])
+    .expect("parse should succeed");
+    assert_eq!(
+        cmd,
+        Command::Container(ContainerArgs {
+            subcommand: ContainerSubcommand::Reset {
+                name: Some("web".to_owned()),
+                keep_data: false,
+                wipe_data: true,
+                yes: true,
             },
             repo_override: None,
             output_json: true,